refactor: extract isEmpty
